Per #1247, moved remove_quotes function to string_manip utility from met_util
